Add PublicAddr for STUNner listener config
This field will be filled in by the operator by finding the service associated with the
listener (containing the listener's name in the annotation
"stunner.l7mp.io/related-gateway-name"). We do not use this field right away, but it will be
useful later for generating ICE server configurations in the WebRTC application server.