Implement new reconciliation logic
The old design mixed creating the reconciliation job queues ("prepare") and actually running the
jobs ("reconciliation"). This caused hangs when during a server restart STUNner was trying to close
a listener that was already reconciled to some different port or protocol. In the new design these
two phases are separated and consecutively, which removed lots of cornercases and hard-to-find
races. In the new design object managers also learned how to create new objects, making the
reconciliation logic fully self-contained.