feature: multi-threaded UDP listeners
This commit adds the ability for STUNner to run UDP listeners over multiple parallel readloops. The
idea is to create a configurable number of UDP server sockets using `SO_REUSEPORT` and spawn a
separate goroutine to run a parallel readloop for each. The kernel will load-balance allocations
across the sockets/readloops per the IP 5-tuple, so the same allocation will always stay at the
same CPU. This allows UDP listeners to scale to multiple CPUs.

Note that this is enabled only for UDP at the moment: TCP, TLS and DTLS listeners spawn a
per-client readloop anyway. Also note that `SO_REUSEPORT` is not portable, so currently we enable
this only for UNIX architectures.

The feature is exposed via the command line flag `--udp-thread-num=<THREAD_NUMBER>` in
`stunnerd`, and the `UDPListenerThreadNum` field of the `Options` struct in the lib.

The commit also adds tests plus a benchmark script that tests `stunnerd` and `turncat` using
iperf. Use `./benchmark.sh udp 16 8000000` to run a benchmark with a UDP listener running 16
readloops, and `./benchmark.sh tcp 0 8000000` for a simple TCP benchmark.