fix: enable health-checking by default
Stunner user to start with health-checking disabled by default and only the gateway operator
enabled health-checking via the stunnerd-config file. This caused problems when stunnerd was
running with no config (e.g., before the user adds a GatewayClass/GatewayConfig), in that it failed
the Kubernetes health-checks. This then caused Kubernetes to mark the stunnerd pods unhealthy and
restart our pods periodically.

This commit sets the default health-checks on. `AdminConfig.HealthCheckEndpoint` is now a pointer:
nil means to use the default health-check URL, pointer to "" means to disable health-checks, and
pointer to an URL means to open the health-check server at the given URL.

Warning: this is a Golang API breaking change (should cause no user visible of the JSON/YAML API
though).