hack: Don't fail readiness checks when there is no config
At the moment after startup STUNner fails readiness checks until it is bootstrapped with a valid
config. Unfortunately, it is a fairly regular case that STUNner runs without a config (e.g., before
the user defines a GatewayClass and/or GatewayConfig) and this causes the kubelet to repeatedly
kill the stunnerd pods.

This commit temporarily changes the readiness behavior of STUNner: it now becomes immediately
"ready" after being started regardless of whether or not there is an actual config, and it remains
so until it receives a SIGTERM. We will revert to the old behavior when the STUNner gateway
operator gains support for managed dataplanes, but until then we need this hack in place for a
reliable operation.