feat: Implement and integrate a Config Discovery Service (CDS) client
This change makes it possible for the STUNner dataplane component `stunnerd` to autodiscover its
own configuration. This is achieved by implementing a CDS client for `stunnerd`: on startup the CDS
client opens a WebSocket connection to a preconfigured CDS server and grabs an initial
configuration. As long as the connection is open, the CDS server will push new configurations down
to the client as they become available. This makes it possible to remove the unstability and
latency associated with obtaining the fresh config by watching a config file mapped into the
filesystem of `stunnerd` (the config file watcher remains available for the "legacy" dataplane
mode).

Major changes:
- implement a generic config watcher interface and make the config file watcher and the CDS config
  poller two implementations of the interface
- obtain the CDS client id from the environment as <STUNNER_NAMESPACE>/<STUNNER_NAME>, or fall back
  to the default (hostname) if the environment is not set
- the initial zeroconfig is now emitted by the config file watcher (the CDS client does not use a
  zeroconfig to init `stunnerd` but rather waits for a valid config from the CDS server)
- the ZeroConfig() and ParseConfig() functions are moved to the CDS client
- stunnerd id is now avaiable in the config (`Admin.Name`)
- tests are implemented at the CDS server side (the operator) that imports the CDS client