fix: Make sure STUNner listens on all interfaces
So far stunnerd has bound TURN server sockets to the corresponding listener address and used the
same listener address to create the relayed transport address. This works well as long as it is
deployed in a Kubernetes pod that has only a single IP address, but breaks on baremetal servers
with multiple IPs. This change causes stunnerd to bind listener sockets to 0.0.0.0:<listener-port>
and only use the listener address to create the relayed transport address.