feat: Implement port range filtering on peer connections
It is now possible to limit the port range available on the peers via STUNner. If your media server
allocates client connections in the [10000:20000] port range, now it is possible to configure
stunnerd to admit peer connections only to that target port range. Denied packets are reported with
an "administratively prohibited error" by the TURN server, and logs are rate-limited to at most 1
report per sec to throttle excessive log message flood.