fix: No longer delay config-delete messages in the CDS server
So far we have added a 5 sec delay before sending our config deletion messages on the CDS server
watchers in order to prevent a race condition where a terminating client receives the config
deletion message from the CDS server before it can actually enter the graceful shutdown cycle,
which causes the immediate dropping of all active client connections. The delay allowed comfortable
time to the client to start graceful shutdown. Unfortunately, the delay also caused unexpected
resets in some hard-to-debug situations, where during adding the Gateway API resources there is a
transient that causes a (delayed) delete due to a missing API resource, which then resets the full
config later when all resources become available.

This change removes the delay in the config-delete path of the CDS server, so all config-delete
updates are immediately sent out. In order to prevent the race condition, the CDS client in
stunnerd ignores delete-config CDS updates all together (but only in stunnerd, the auth-service and
stunnerctl still process them): once started, there is no other way to stop a stunnerd pod than to
delete the Gateway resource.