doc: Fix stunner namespace in the magic mirror example
Without this, stunner was installed in the default namespace, while the
configuration file "kurento-magic-mirror-stunner.yaml" expects it to be
in the "stunner" namespace.

This caused 'namespace "stunner" not found' errors when following the
demo steps and then configuring stunner by applying
"kurento-magic-mirror-stunner.yaml".

By installing stunner in the "stunner" namespace, the demo installation
works as expected.