fix: Fix ready status reporting, fix #201
Formerly, stunnerd used to report a "ready" status right after initialization. This meant that
stunnerd was receiving live traffic even before obtaining a valid configuration. This fix changes
this behavior: we now only report ready status if (1) we are not being shut down, (2) we have
finished successfully reconciling the first valid (non-empty, ie., non-bootstrap config), and (3)
the reconcile is not due to a rollback to an earlier config. From that point stunnerd reports ready
status until being shut down, even when losing the connection to the operator (this helps a cluster
stay alive while the operator restarts).